Create is_cell_empty? method
The is_cell_empty? method verifies given a row and column indices if
that cell is empty, feature that will be useful when the player wants
to place a mark on a non empty cell.

Also fixed a typo at the board variable and added default getter and
setter methods for that variable.