Create get_coordinates method and game function
The Board class now has a COORDINATES hash that it's a constant; it
has some directions as keys that represent the positions on the board
and a list with the row and column as values. This is used in the new
get_coordinates method.

The get_coordinates method receives a direction and then tries to get
one value from the COORDINATES hash. In case the value is not found
it returns the "unknown" string.

The game function will play a tic-tac-toe game. For the moment it only
creates two players, one with its mark; then it displays the board at
the start of the game and after the first round. Also it keeps asking
the player the position of the mark until it gives a valid direction.