Change game function
First, added a new to_s method to Player's class; this one is to display
the player's mark during each round.

In the game function, now there is a loop of 9 iterations where it first
goes the "X" player, then goes the "O" player until one of the two win
the game (this is still not working) and the previous code of one turn
is inside this loop to simulate the game. When this loop ends, it means
that the game ended in a tie.