Finish game function
To finish this function it is needed a way to identify a winner; to do
that the Board function now has two new methods: the update_lines and
the is_line_filled? methods.

The update_lines it just creates and return a new class variable: a hash
called lines, which contains a series of posible ways to win the game.
The keys of this hash is a way to easily recognize the line with a
string; the values are lists, one for each element that contains the
uptated values in the game.

The is_line_filled? method, like its name implies, verifies if a user
has filled a line, thanks to the previous method to check the lines of
the board. It iterates through that hash and when it founds that one
line has all the same value, it returns true. At the end it returns
false.

Finally, in the game function, after the player has placed a mark, it
checks with the methods named above if the current player won the game;
in case this happens, it shows the board one last time and announces the
winner. In the tie it also shows the board and announces the result of
the game.