Add validate_number method
This method has one parameter: input. It checks if the input is a
number, this number is between 1 and 9 and finally if the cell that it
refers to is empty. When all this conditions are met, it returns true.
It returns false if one of the three is not true.