Add check_diagonal method
It has four parameters: row, column, player_mark and direction. First,
it gets the diagonal coordinates with the diagonal_coordinates method.
Then it returns if the coordinates_in_line? method returns false. When
that method returns true, it gets the diagonal values with the
diagonal_values method. Finally, it returns true if the line_filled?
method returns true; otherwise it returns nil.