Add winner? method
It has three parameters: row, column and player_mark. First, it gets the
row values with the obtain_row method. Then it returns true if the
line_filled? method returns true. When this method returns false, then
it gets the column values with the obtain_column method. Now it returns
true if the line_filled? method returns true. When this method returns
false, now it returns true if the check_diagonal method returns true.
When it returns nil, it now returns true if the check_diagonal method
returns true. When this method returns nil, now it returns false.